organize RemoteGit to highlight interfaces (refs #24)
