+++ src/MBO/RemoteGit/Filter/ProjectTypeFilter.php
@@ -2,9 +2,12 @@
 
 namespace MBO\RemoteGit\Filter;
 
-use MBO\RemoteGit\ProjectInterface;
-use MBO\RemoteGit\ClientInterface as GitClientInterface;
 use Psr\Log\LoggerInterface;
+
+use MBO\RemoteGit\ProjectInterface;
+use MBO\RemoteGit\ProjectFilterInterface;
+use MBO\RemoteGit\ClientInterface as GitClientInterface;
+
 
 /**
  * Ignore project according to their type
