+++ src/MBO/RemoteGit/Gitlab/GitlabClient.php
@@ -1,11 +1,15 @@
 <?php
 
-namespace MBO\RemoteGit;
+namespace MBO\RemoteGit\Gitlab;
 
-use \GuzzleHttp\Client as GuzzleHttpClient;
 use Psr\Log\LoggerInterface;
+use \GuzzleHttp\Client as GuzzleHttpClient;
 
-use MBO\RemoteGit\Filter\ProjectFilterInterface;
+use MBO\RemoteGit\ClientInterface;
+use MBO\RemoteGit\ProjectInterface;
+use MBO\RemoteGit\FindOptions;
+use MBO\RemoteGit\ProjectFilterInterface;
+
 
 /**
  * Find gitlab projects
@@ -57,13 +61,13 @@ class GitlabClient implements ClientInterface {
         foreach ( $options->getUsers() as $user ){
             $result = array_merge($result,$this->findByUser(
                 $user,
-                $options->getFilterCollection()
+                $options->getFilter()
             ));
         }
         foreach ( $options->getOrganizations() as $org ){
             $result = array_merge($result,$this->findByGroup(
                 $org,
-                $options->getFilterCollection()
+                $options->getFilter()
             ));
         }
         return $result;
@@ -113,7 +117,7 @@ class GitlabClient implements ClientInterface {
         return $this->fetchAllPages(
             $path,
             $params,
-            $options->getFilterCollection()
+            $options->getFilter()
         );
     }
